temporarily disable ArchiveItAccw functionality
The ArchiveItAccw collection data is loaded directly from the Archive
It website: https://archive-it.org.  Due to a massive power outage in
San Francisco on 2014-12-11, the Internet Archive sites, including
Archive It, went offline.  This outage breaks the ArchiveItAccw data
loading step in the Ichabod cucumber setup.

This branch temporarily disables all ArchiveItAccw functionality until
the Internet Archive websites are back up and a cassette can be
recorded for ArchiveItAccw cucumber scenarios.