update ArchiveItAccw scenarios
The ArchiveItAccw scenarios relied on data taken from the live site:
https://archive-it.org/collections/4049.json

The scenarios broke when the URL for a site being harvested as part
of the Archive It service changed from http://LouisKarchin.com
to http://www.louiskarchin.com.

An MD5 hash of the URL is used to generate the Ichabod identifier.
So, the change in the website URL meant that the identifiers described
in the scenarios was no longer present, thus the scenarios failed.

Similarly the search.feature searched for "louiskarchin.com" and
not "www.louiskarchin.com", so that scenario failed as well.

The scenarios have been fixed, and the data loading changed to
use a capture of part of the ArchiveIt site instead of the live site.